requirements.txt
aiohttp==3.8.3
beautifulsoup4==4.11.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.126.0
lxml==4.9.2